pyproject.toml
[tool.poetry]
name = "nonebot_plugin_justsix"
version = "1.0.0"
description = "一个nonebot2的发6插件"
authors = ["Maximilian Wu <me@maxng.cc>"]
repository = "https://github.com/iidamie/nonebot_plugin_justsix"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = ">=2.0.0-beta.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
